Added manipVirtuals (Dict) field in RenderObject, intent is to represent standard virtual function for manip from subscreen mgr
Intent is that each RenderObject may export the set of functions to be used for object manipulation in its representation. The
virtual interface will have a fixed specification close to OpenGL definitions (model view, projections, viewport etc...). The
upshot is if you want to rotate in model space, you will look for the rotateFn (name may still vary), if it exists then it
will provide the rotation, if it does not exist, then the object has not been designed with this capability.