Some cleanup + got rid of GeometryTypes
For the cleanup, it's basically putting stuff in order in buffer.jl and texture.jl. Namely, constructors first below the struct def, and functionality related to  the struct after the struct. 

In getting rid of GeometryTypes there were two functions in for_moderngl.jl that were deleted and I also rewrote glsl_alignment_size in uniformbuffer.jl to hopefully work with generic Matrices and Vectors => Please check if this is correct (also the documentation) 

I also changed GLBuffer to Buffer and GLContext to Context, and removed all the GLAbstraction.'s